inted test script

+++ tests/testLibrosaFeature.py
@@ -90,7 +90,7 @@ def test_melfb():
 
         # Our version only returns the real-valued part.
         # Pad out.
-        wts = numpy.pad(wts, [ (0, 0), (0, DATA['nfft'][0]/2 - 1)], mode='constant')
+        wts = numpy.pad(wts, [ (0, 0), (0, int(DATA['nfft'][0]/2 - 1))], mode='constant')
                                 
         assert wts.shape == DATA['wts'].shape
